requirements.txt
langchain>=0.1.0
langgraph>=0.0.30
langchain-aws>=0.1.0
boto3>=1.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
